📝 docs: map 명세서 작성
